requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions

rtree==0.9.7
shapely==1.7.1